Rename solution-picker to solution_picker per snake_case style.
