+++ lib/MrMurano/commands/solution_picker.rb
